Update cnn_tensor_flow.py
Commented out some code to run the test images through the model instead of the original train/test split.  Will have to go back and adjust some of the code to get rid of the train/test split of the original train data.  Initial run of the model results in near 80/20% predictions for the two classes, which is the exact split of the classes in the training set.  The initial predictions suggest the model is just guessing and is not actually learning anything.